Time: 4 ms (31.35%) | Memory: 8.6 MB (37.89%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -5,7 +5,7 @@ public:
         int num=x;
         long long rev=0;
         while(x>0){
-                long long rem=x%10;
+                int rem=x%10;
                 rev=rev*10+rem;
                 x=x/10;
         }
